Make sure that the text view puts the cursor at the end when it first becomes first responder
When the text view is initialized, it is empty, and so its selected range is (0,0).
After it’s populated with text, it would normally put the cursor at the end of the
text, except for https://github.com/kenmueller/TextView/pull/21 which reset
the position to the beginning.

We switch from using `selectedRange` to `selectedTextRange` because
we can set the latter to the `endOfDocument` position.